Fixing editor Camera and ImGuizmo's bugs...

+++ HazleEngine/src/Hazle/Scene/SceneSerializer.cpp
@@ -183,7 +183,7 @@ namespace Hazle
 		{
 			for (auto entity : entities)
 			{
-				uint32_t uuid = entity["Entity"].as<uint64_t>();
+				uint32_t uuid = entity["Entity"].as<uint32_t>();
 
 				std::string name;
 				auto tagComponent = entity["TagComponent"];
